backend/requirements.txt
# Requires Python 3.13+ (3.14 not yet supported by pydantic-core)
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
apscheduler==3.10.4
python-nmap==0.7.1
pyyaml==6.0.2
websockets==13.1
httpx==0.27.2

# Dev
ruff==0.6.9
mypy==1.11.2
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
